fix: address review comments for MongoDB monitor
- Normalize nil database_query to default value "{\"ping\": 1}" to
  prevent spurious diffs in the Read method, matching MySQL monitor behavior
- Extract reusable strToPtr and ptrToTypes helper functions to shared
  resource_monitor_helpers.go file for better maintainability
- Add TestAccMonitorMongoDBResourceImport test to verify import
  functionality, ensuring resources can be imported by ID

Amp-Thread-ID: https://ampcode.com/threads/T-019c0da7-4cc1-724b-8ef5-ab0b38faebab
